pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "botds-openai"
version = "0.1.0"
description = "Bot Data Scientist with OpenAI-led decisions"
authors = [{name = "Bot DS Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7",
    "pyyaml",
    "numpy",
    "pandas", 
    "scikit-learn",
    "matplotlib",
    "tqdm",
    "requests",
    "python-dotenv",
    "joblib",
    "openai>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "isort",
    "mypy"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["botds*", "cli*"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
